Consolidated core interfaces.
