--- src/renci_ner/services/core.py
@@ -1,38 +1,0 @@
-
-from renci_ner.annotations import AnnotationProvenance
-
-
-class Annotator:
-    """
-    An interface for annotating text using a service.
-    """
-
-    @property
-    def provenance(self): # -> "AnnotationProvenance":
-        """
-        Return an AnnotationProvenance describing annotations produced by this service.
-
-        :return AnnotationProvenance: The provenance of annotations generated by this Annotator.:
-        """
-        return AnnotationProvenance(
-            name="Annotator",
-            url="http://example.org/",
-            version="0.0.1",
-        )
-
-    def annotate(self, text: str, props: dict = {}): # -> "AnnotatedText":
-        """
-        Annotate a text with a set of service-specific properties.
-
-        :return AnnotatedText: The annotated text.
-        """
-        pass
-
-    def supported_properties(self) -> dict[str, str]:
-        """
-        Return a list of supported properties for this service.
-
-        :return: A dictionary of supported properties, with the values describing each property.
-        """
-        return {}
-
