+++ tests/sapbert/test_sapbert.py
@@ -1,4 +1,4 @@
-from renci_ner.annotations import AnnotationProvenance
+from renci_ner.core import AnnotationProvenance
 from renci_ner.services.linkers.sapbert import SAPBERTAnnotator
